# swagger.yaml
openapi: 3.0.0
info:
  title: UserHub API
  version: 1.0.0
  description: |
    📘 **Documentació de l'API UserHub**

    Aquesta API senzilla permet gestionar usuaris amb operacions CRUD:

    - 🔍 Llistar usuaris (`GET /usuaris`)
    - ➕ Crear un usuari (`POST /usuaris`)
    - ✏️ Actualitzar un usuari (`PUT /usuaris/{id}`)
    - ❌ Esborrar un usuari (`DELETE /usuaris/{id}`)

    També inclou una ruta bàsica (`GET /`) que mostra un missatge HTML de benvinguda.

    👉 Consulta cada endpoint per veure exemples i paràmetres disponibles.
servers:
  - url: http://localhost:3000
    description: Servidor local
paths:
  /:
    get:
      summary: Pàgina principal
      description: |
        Ruta bàsica. 
        Mostra un missatge HTML amb el títol de l'aplicació.

        Exemple URL: [/](http://localhost:3000/)
      responses:
        '200':
          description: HTML amb títol de l'aplicació.
          content:
            text/html:
              schema:
                type: string
                example: <h1>APLICACIÓ PER LA GESTIÓ D'USUARIS</h1>
  /usuaris:
    get:
      summary: Obté una llista d'usuaris
      description: |
        GET /usuaris - Retorna tots els usuaris o els filtra per nom i edat amb paràmetres de query.

        Exemple URLs:
        [/usuaris](http://localhost:3000/usuaris)
        [/usuaris?nom=Maria&edat=25](http://localhost:3000/usuaris?nom=Maria&edat=25)
      parameters:
        - in: query
          name: nom
          schema:
            type: string
          description: Filtra per nom de l'usuari
        - in: query
          name: edat
          schema:
            type: integer
          description: Filtra per edat de l'usuari
      responses:
        '200':
          description: Llista d'usuaris trobats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuari'
        '404':
          description: Cap usuari trobat
    post:
      summary: Crea un nou usuari
      description: |
        POST /usuaris - Afegeix un nou usuari amb nom i edat.

        Exemple JSON:
        {
          "nom": "Ramon",
          "edat": 50
        }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuariInput'
      responses:
        '201':
          description: Usuari creat correctament
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuari'
  /usuaris/{id}:
    get:
      summary: Obté un usuari per ID
      description: |
        GET /usuaris/:id - Retorna l'usuari amb l'ID especificat via paràmetres de ruta.

        Exemple URL: [/usuaris/2](http://localhost:3000/usuaris/2)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de l'usuari
      responses:
        '200':
          description: Usuari trobat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuari'
        '404':
          description: Usuari no trobat
    put:
      summary: Actualitza un usuari per ID
      description: |
        PUT /usuaris/:id - Actualitza el nom i l'edat de l'usuari amb l'ID especificat.

        Exemple JSON:
        {
          "edat": 100
        }
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de l'usuari
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuariInput'
      responses:
        '200':
          description: Usuari actualitzat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuari'
        '404':
          description: Usuari no trobat
    delete:
      summary: Esborra un usuari per ID
      description: |
        DELETE /usuaris/:id - Esborra l'usuari amb l'ID especificat.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de l'usuari
      responses:
        '200':
          description: Usuari esborrat correctament
          content:
            text/plain:
              schema:
                type: string
                example: Usuari 2 esborrat
        '404':
          description: Usuari no trobat
components:
  schemas:
    Usuari:
      type: object
      properties:
        id:
          type: integer
          example: 2
        nom:
          type: string
          example: "Maria"
        edat:
          type: integer
          example: 25
    UsuariInput:
      type: object
      required:
        - nom
        - edat
      properties:
        nom:
          type: string
          example: "Ramon"
        edat:
          type: integer
          example: 50